implement favorite photo #3 - fix issue doesn't update favorite button in paging item

+++ app/src/main/java/com/psd/learn/mysplash/ui/feed/PagingFeedViewModel.kt
@@ -1,6 +1,5 @@
 package com.psd.learn.mysplash.ui.feed
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
@@ -10,10 +9,10 @@ import com.psd.learn.mysplash.data.local.datasource.PhotosLocalRepository
 import com.psd.learn.mysplash.data.local.entity.CollectionItem
 import com.psd.learn.mysplash.data.local.entity.PhotoItem
 import com.psd.learn.mysplash.data.remote.repository.UnSplashPagingRepository
-import com.psd.learn.mysplash.utils.log.Logger
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
@@ -26,6 +25,8 @@ class PagingFeedViewModel @Inject constructor(
     private val photosLocalRepo: PhotosLocalRepository
 ) : ViewModel() {
 
+    private val favoriteActionStateFlow = MutableStateFlow<List<FavoriteAction>>(emptyList())
+
     val collectionPagingDataFlow: Flow<PagingData<CollectionItem>> = pagingRepository
         .getFeedCollectionsStream()
         .cachedIn(viewModelScope)
@@ -36,31 +37,65 @@ class PagingFeedViewModel @Inject constructor(
 
     val photoPagingDataFlow: Flow<PagingData<PhotoItem>> = pagingRepository
         .getFeedPhotosStream()
+        .map { pagingData: PagingData<PhotoItem> ->
+            pagingData.map {
+                val isFavorite = photosLocalRepo.checkFavoritePhotoById(it.photoId)
+                if (isFavorite) it.copy(isFavorite = true)
+                else it
+            }
+        }
+        .flowOn(Dispatchers.IO)
         .cachedIn(viewModelScope)
-        .combine(photosLocalRepo.getPhotoIdsStream()) { remotePagingPhotos, localPhotoIds ->
-            remotePagingPhotos.map { photoItem ->
-                val isFavorite = localPhotoIds.contains(photoItem.photoId)
-                Logger.d("sangpd", "PagingFeedViewModel: isFavorite: $isFavorite")
-                if (isFavorite) {
-                    photoItem.copy(isFavorite = true)
-                } else {
-                    photoItem
+        .combine(favoriteActionStateFlow) { pagingData, actions: List<FavoriteAction> ->
+            actions.fold(pagingData) { acc: PagingData<PhotoItem>, event: FavoriteAction ->
+                applyEvent(acc, event)
+            }
+        }
+
+    fun addOrRemoveFavorite(photoItem: PhotoItem, status: Boolean) {
+        viewModelScope.launch {
+            val action: FavoriteAction
+            if (status) {
+                action = FavoriteAction.RemoveFavorite(photoItem)
+                photosLocalRepo.removeFavoritePhoto(photoItem)
+            } else {
+                action = FavoriteAction.AddFavorite(photoItem)
+                photosLocalRepo.addFavoritePhoto(photoItem.copy(isFavorite = true))
+            }
+            onFavoriteAction(action)
+        }
+    }
+
+
+    fun onFavoriteAction(action: FavoriteAction) {
+        favoriteActionStateFlow.value += action
+    }
+    
+    private fun applyEvent(
+        pagingData: PagingData<PhotoItem>,
+        action: FavoriteAction
+    ) : PagingData<PhotoItem> {
+        return when(action) {
+            is FavoriteAction.AddFavorite -> {
+                pagingData.map {
+                    if (action.data.photoId == it.photoId) return@map it.copy(isFavorite = true)
+                    else return@map it
+
+                }
+            }
+            is FavoriteAction.RemoveFavorite -> {
+                pagingData.map {
+                    if (action.data.photoId == it.photoId) return@map it.copy(isFavorite = false)
+                    else return@map it
                 }
             }
         }
-        .cachedIn(viewModelScope)
-
-    fun insertFavoritePhoto(photoItem: PhotoItem) {
-        viewModelScope.launch {
-            photosLocalRepo.addFavoritePhoto(photoItem)
-        }
     }
 
-    fun removeFavoritePhoto(photoItem: PhotoItem) {
-        viewModelScope.launch {
-            photosLocalRepo.removeFavoritePhoto(photoItem)
-        }
-    }
+}
 
+sealed class FavoriteAction {
+    data class AddFavorite(val data: PhotoItem) : FavoriteAction()
+    data class RemoveFavorite(val data: PhotoItem) : FavoriteAction()
 }
